sepreated class
